Conflicts:
	application/controllers/admin/tokens.php
	application/views/admin/token/browse.php

+++ application/controllers/admin/surveyadmin.php
@@ -490,7 +490,7 @@ class SurveyAdmin extends Survey_Common_Action
         if (empty(Yii::app()->session['USER_RIGHT_SUPERADMIN']))
             $surveys->permission(Yii::app()->user->getId());
         $surveys = $surveys->with('languagesettings', 'owner')->findAll();
-
+        $aSurveyEntries = new stdClass();
         $aSurveyEntries->page = 1;
         foreach ($surveys as $rows)
         {
@@ -924,7 +924,7 @@ class SurveyAdmin extends Survey_Common_Action
             {
                 $importerror = true;
             }
-            if ($action == 'importsurvey')
+            if ($action == 'importsurvey' && isset($sFullFilepath))
             {
                 unlink($sFullFilepath);
             }
@@ -1095,7 +1095,7 @@ class SurveyAdmin extends Survey_Common_Action
                     $esresult['template']=Yii::app()->getConfig('defaulttemplate');
                 }
                 $esresult['template']=validateTemplateDir($esresult['template']);
-                
+
                 $esrow = $esresult;
             }
         }
